Reworked the discovery process to be faster and more flexible
Added a proper browser class that can be used to better control how
devices are discovered. It allows for both blocking behaviour where it
waits for a set of devices in a time-span that matches a filter
(typically for scripts), or it can be used to listen for devices as
long as the process is running (for daemons).

The get_chromecast* functions will create this browser on-demand and
will work as they used to before this chane.

The Chromecast class has been moved from the main package and into a
'device' module.

If you are looking for unique values for a Chromecast, e.g. the UUID
or IP the discovery system will now return as soon as that device
appears, it will no longer wait for the timeout to occur.
This is handled by using a blocking queue on the main thread, as soon
as a device is discovered it will placed in the queue and the thread
wakes up, if the criteria is met it returns immediately.

In addition the Chromecast class can now be created without connecting
to the device via the socket client. If the 'connect' parameter is
False it will not connect and leave it up to the caller to call
connect() when required.
The default is to connect to filtered devices before returning them but
this may also be turned off. The exception is if filters that require a
connection to work (e.g. status.app_id), if used it will connect to the
device as it finds them to allow the filter to work.